Implement HARD_RESET exchange.
This is kind of drafty in the sense that the code hasn't run before,
but at least it's clearly defining what to do and how to parse the
incoming packets, as well as defining how to generate outgoing
packets.

The next step is to make the tests able to edit the openvpn server
config so that they can set no-auth, and then implement a first test
that validates that the HARD_RESET exchange is properly done.