More debugging and... TCP?!
Just found out that openvpn has a tcp mode.

It kind of changes everything.

So far I thought that I _had_ to use UDP anyway, so I might as well go
for something generic. But if TCP is supported, then it means it
unlocks even more locked down environments, so I am switching the
project to be an openvpn client in TCP mode. Hopefully it'll support
more (it looks like some VPNs like SoftEther support some TCP-based
mode as well?) by default, but I honestly don't expect it. It's
basically becoming an exclusively openvpn client now.

What's great about it is that I can extensively test all modes and be
certain that my code is correct in all ways now.

What's less great is that I need to rewrite all the markdown files and
redo a bunch of code.